requirements.txt

streamlit==1.53.1
python-dotenv==1.2.1
google-generativeai==0.8.6
pillow==12.1.0